+++ src/cljs/rems/administration/administration.cljs
@@ -16,54 +16,57 @@
  (fn [db _]
    (::loading? db)))
 
-(defn- to-administration [class]
-  [atoms/link {:class class} "/administration" (text :t.navigation/administration)])
+(defn- nav-classes [active?]
+  (into [:nav-item :nav-link]
+        (when active?
+          [:active])))
 
-(defn- to-catalogue-items [class]
-  [atoms/link {:class class} "/administration/catalogue-items" (text :t.administration/catalogue-items)])
+(defn- to-administration [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration" (text :t.navigation/administration)])
 
-(defn- to-resources [class]
-  [atoms/link {:class class} "/administration/resources" (text :t.administration/resources)])
+(defn- to-catalogue-items [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration/catalogue-items" (text :t.administration/catalogue-items)])
 
-(defn- to-forms [class]
-  [atoms/link {:class class} "/administration/forms" (text :t.administration/forms)])
+(defn- to-resources [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration/resources" (text :t.administration/resources)])
 
-(defn- to-workflows [class]
-  [atoms/link {:class class} "/administration/workflows" (text :t.administration/workflows)])
+(defn- to-forms [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration/forms" (text :t.administration/forms)])
 
-(defn- to-licenses [class]
-  [atoms/link {:class class} "/administration/licenses" (text :t.administration/licenses)])
+(defn- to-workflows [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration/workflows" (text :t.administration/workflows)])
+
+(defn- to-licenses [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration/licenses" (text :t.administration/licenses)])
+
+(defn- to-blacklist [active?]
+  [atoms/link {:class (nav-classes active?)} "/administration/blacklist" (text :t.administration/blacklist)])
 
 (defn administration-navigator [selected]
   [:div.navbar.mb-4.mr-auto.ml-auto
-   [to-administration [:nav-item :nav-link (when (contains? #{:rems.administration/administration}
-                                                            selected)
-                                             :active)]]
-   [to-catalogue-items [:nav-item :nav-link (when (contains? #{:rems.administration/catalogue-items
-                                                               :rems.administration/catalogue-item
-                                                               :rems.administration/create-catalogue-item}
-                                                             selected)
-                                              :active)]]
-   [to-resources [:nav-item :nav-link (when (contains? #{:rems.administration/resources
-                                                         :rems.administration/resource
-                                                         :rems.administration/create-resource}
-                                                       selected)
-                                        :active)]]
-   [to-forms [:nav-item :nav-link (when (contains? #{:rems.administration/forms
-                                                     :rems.administration/form
-                                                     :rems.administration/create-form}
-                                                   selected)
-                                    :active)]]
-   [to-workflows [:nav-item :nav-link (when (contains? #{:rems.administration/workflows
-                                                         :rems.administration/workflow
-                                                         :rems.administration/create-workflow}
-                                                       selected)
-                                        :active)]]
-   [to-licenses [:nav-item :nav-link (when (contains? #{:rems.administration/licenses
-                                                        :rems.administration/license
-                                                        :rems.administration/create-license}
-                                                      selected)
-                                       :active)]]])
+   [to-administration (contains? #{:rems.administration/administration} selected)]
+   [to-catalogue-items (contains? #{:rems.administration/catalogue-items
+                                    :rems.administration/catalogue-item
+                                    :rems.administration/create-catalogue-item}
+                                  selected)]
+   [to-resources (contains? #{:rems.administration/resources
+                              :rems.administration/resource
+                              :rems.administration/create-resource}
+                            selected)]
+   [to-forms (contains? #{:rems.administration/forms
+                          :rems.administration/form
+                          :rems.administration/create-form}
+                        selected)]
+   [to-workflows (contains? #{:rems.administration/workflows
+                              :rems.administration/workflow
+                              :rems.administration/create-workflow}
+                            selected)]
+   [to-licenses (contains? #{:rems.administration/licenses
+                             :rems.administration/license
+                             :rems.administration/create-license}
+                           selected)]
+   [to-blacklist (contains? #{:rems.administration/blacklist}
+                            selected)]])
 
 (defn administration-navigator-container
   "Component for showing a navigator in the administration pages.